src/requirements.txt
pillow
tk
customtkinter
requests
